add IsDuplicated
IsDuplicated determines whether an interger array contains any duplicated values

+++ ArraysAndStrings/IsUnique.go
@@ -1,17 +1,18 @@
 package arraysandstrings
 
+//IsUnique determines whether a string contains duplicate characters
 func IsUnique(str string) bool {
 	// ASCII consists of 128 characters
 	if len(str) > 128 {
 		return false
 	}
-	characterSet := make(map[int]bool, 128)
+	characterSet := make(map[byte]bool, 128)
 
 	for _, character := range []byte(str) {
-		if characterSet[int(character)] {
+		if characterSet[character] {
 			return false
 		}
-		characterSet[int(character)] = true
+		characterSet[character] = true
 	}
 	return true
 }
